requirements.txt
django==3.2.2
djangorestframework==3.12.4
mysqlclient==2.0.1
